Import libsoup-2.50.0 as net/libsoup, replaces net/libsoup24.
Libsoup is an HTTP library implementation in C. It was originally part
of a SOAP (Simple Object Access Protocol) implementation called Soup, but
the SOAP and non-SOAP parts have now been split into separate packages.

libsoup uses the Glib main loop and is designed to work well with GTK
applications. This enables GNOME applications to access HTTP servers
on the network in a completely asynchronous fashion, very similar to
the Gtk+ programming model (a synchronous operation mode is also
supported for those who want it).

Features:
  * Completely Asynchronous
  * Connection cache
  * HTTP chunked transfer support
  * HTTP, SOCKS4, and SOCKS5 authenticated proxy support
  * SSL Support using OpenSSL or GnuTLS
  * Client support for Digest, NTLM, and Basic authentication
  * HTTP server
  * Server support for Digest and Basic authentication