Remove reference to file no longer part of the perl distribution.
Add a Configure test to verify that including <fenv.h> doesn't produce
a build error, as it will in quite few cases on NetBSD on archs which
are not amd64, i386 or sparc in NetBSD 6.x.  If the test build fails,
pretend we don't have fenv.h.

Validated that the result builds on NetBSD/evbarm 6.0 and NetBSD/i386 6.1.5.

Build fix, so no need to bump PKGREVISION.

OK by wiz@