Import lbxproxy-1.0.3 as x11/lbxproxy.
Applications that would like to take advantage of the Low Bandwidth extension to
X (LBX) must make their connections to an lbxproxy. These applications need to
know nothing about LBX, they simply connect to the lbxproxy as if were a regular
server. The lbxproxy accepts client connections, multiplexes them over a single
connection to the X server, and performs various optimizations on the X protocol
to make it faster over low bandwidth and/or high latency connections.

With regard to authentication/authorization, lbxproxy simply passes along to the
server the credentials presented by the client. Since X clients will connect to
lbxproxy, it is important that the user's .Xauthority file contain entries with
valid keys associated with the network ID of the proxy. lbxproxy does not get
involved with how these entries are added to the .Xauthority file. The user is
responsible for setting it up.

Note that the X server source from X.Org no longer supports the LBX extension,
so this program is only useful in connecting to older X servers.