Update tex-minitoc{,-doc} to 61
The present release has no functional changes.  Its only purpose was to
reduce the size of the documentation, from 25+mb to <2mb, by removing
images not important to the content.  This work was done by Nils Ole
Tippenhauer, and the release was constructed by Karl Berry, but neither
of us are prospective maintainers of minitoc.