use -fno-reorder-blocks for sparc64, mips, and vax and GCC 4.5*.
something in op.c (as miniop.c) is mis-compiled with this option which
is enabled by -O2, when using GCC 4.5.  i didn't try to figure out
exactly what as op.c is 419,359 bytes long and the assembler output
is almost 100% different and approximiately 1.5MB either way (the
diff of the asm output is larger than the combined inputs), so for now
we have this hack.  this problem doesn't appear to occur in newer GCC.


XXX: pullup to 2015Q2.