Update to 0.32.8:
A new pixman release 0.32.8 is now available. This is a stable release that
contains an important bug fix (buffer overflow), which can affect 32-bit
systems.

I recommend to all the distributions that still use the 0.32.x version to
upgrade to this release as soon as possible.

For those distributions that already moved to the new 0.33.2 version, I will
release in the coming weeks a new version (0.33.4), which will include this
bug fix, among other things.